requirements.txt
Flask==1.0.2
numpy==1.15.1
pandas==0.23.4
gunicorn==19.9.0
dash==0.27.0
dash-table-experiments==0.6.0
dash_html_components==0.13.2
dash_core_components==0.30.2
python-dotenv-0.9.1
tenacity==5.0.2
sqlalchemy==1.2.12
psycopg2==2.7.5
typing
pytest==3.8.1
APScheduler==3.0.0